preprocess/pyproject.toml
[project]
name = "hlab-preprocess"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2","numpy>=1.26","opencv-python-headless>=4.9",
  "albumentations>=1.4","pyyaml>=6.0"
]
